Add file names sort for verification step
On rotational HDDs it could make sense for faster sequential reading since the data on copy (or after defragmentation) is usually placed block-by-block sequentially.
While on SSDs it would not have any impact (except sort itself) at all.

On my tests on rotational HDDs it was:
- 3-4x faster verification in terms of files processed per second: from 4-5 fps to 15-20 fps
- 2x faster throughput: from 70-80MBps to 130-150MBps
- caching become possible for OS (Linux or macOS now should cache the data much better which could make faster re-verification if canceled)

So assume it makes sense, probably later the switch to disable it will be added to do not waste time
on sort while the data is on SSD.

Anyway more tests is required so leaving it as a default behaviour for that purpose.

Signed-off-by: Igor Shishkin <me@teran.ru>